Move blosc files into their own subdirectory. Move src/ and include/ to this directory. Add cmake/ directory. Add CMakeLists.txt to encapsulate feature detection for avx2 and sse2. The blosc sources have #error if an enabling symbol is not defined, so we detect that. Added two detection sources to cmake directory to enable this. Turn off compilation warnings for blosc files.
