+++ external/blosc/LICENSE
@@ -2,7 +2,8 @@ BSD License
 
 For Blosc - A blocking, shuffling and lossless compression library
 
-Copyright (C) 2009-2017 Francesc Alted <francesc@blosc.org>
+Copyright (C) 2009-2018 Francesc Alted <francesc@blosc.org>
+Copyright (C) 2019-present Blosc Development team <blosc@blosc.org>
 
 Redistribution and use in source and binary forms, with or without modification,
 are permitted provided that the following conditions are met:
