+++ external/blosc/include/bitshuffle-generic.h
