+++ external/blosc/include/shuffle-avx2.h
