+++ external/blosc/include/shuffle-generic.h
