+++ external/blosc/include/shuffle.h
