+++ external/blosc/src/shuffle-avx2.c
