+++ external/blosc/src/shuffle-generic.c
