+++ external/blosc/src/shuffle.c
